+++ src/UI/SignTransactionsModals/SignWithDeviceModal/components/SignStepBody.tsx
@@ -1,26 +1,26 @@
 import React from 'react';
-import { faExclamationTriangle } from '@fortawesome/free-solid-svg-icons';
-import { FontAwesomeIcon } from '@fortawesome/react-fontawesome';
 import { Address } from '@multiversx/sdk-core/out';
 import classNames from 'classnames';
 
-import globalStyles from 'assets/sass/main.scss';
-import { useGetNetworkConfig } from 'hooks';
+import { useGetEgldPrice, useGetNetworkConfig } from 'hooks';
 import { useGetTokenDetails } from 'hooks/transactions/useGetTokenDetails';
 import type {
   ActiveLedgerTransactionType,
   MultiSignTransactionType
 } from 'types';
-import { ProgressSteps } from 'UI/ProgressSteps';
-import { TokenDetails } from 'UI/TokenDetails';
+import { NftEnumType } from 'types/tokens.types';
 import { TransactionData } from 'UI/TransactionData';
-
 import { getIdentifierType } from 'utils';
 import { getEgldLabel } from 'utils/network/getEgldLabel';
 import { formatAmount } from 'utils/operations/formatAmount';
 import { isTokenTransfer } from 'utils/transactions/isTokenTransfer';
 
 import { useSignStepsClasses } from '../hooks';
+import { ConfirmAmount } from './components/ConfirmAmount';
+import { ConfirmFee } from './components/ConfirmFee';
+import { ConfirmReceiver } from './components/ConfirmReceiver';
+import { NftSftPreviewComponent } from './components/NftSftPreviewComponent';
+import styles from './signStepBodyStyles.scss';
 
 export interface SignStepInnerClassesType {
   buttonsWrapperClassName?: string;
@@ -46,9 +46,6 @@ export interface SignStepBodyPropsType {
 export const SignStepBody = ({
   currentTransaction,
   error,
-  allTransactions,
-  isGuarded,
-  currentStep,
   signStepInnerClasses
 }: SignStepBodyPropsType) => {
   const egldLabel = getEgldLabel();
@@ -58,14 +55,11 @@ export const SignStepBody = ({
   }
 
   const { network } = useGetNetworkConfig();
-
   const {
     inputGroupClassName,
     inputLabelClassName,
     inputValueClassName,
-    errorClassName,
-    scamAlertClassName,
-    progressClassName
+    errorClassName
   } = signStepInnerClasses || {};
 
   const { tokenId, nonce, amount, multiTxData, receiver } =
@@ -75,15 +69,16 @@ export const SignStepBody = ({
     tokenId && isTokenTransfer({ tokenId, erdLabel: egldLabel })
   );
 
-  const { isNft } = getIdentifierType(tokenId);
+  const { isNft, isEgld, isEsdt } = getIdentifierType(tokenId);
 
   // If the token has a nonce means that this is an NFT. Eg: TokenId=TOKEN-1hfr, nonce=123 => NFT id=TOKEN-1hfr-123
   const appendedNonce = nonce ? `-${nonce}` : '';
   const nftId = `${tokenId}${appendedNonce}`;
 
-  const { tokenDecimals, tokenAvatar } = useGetTokenDetails({
-    tokenId: nonce && nonce.length > 0 ? nftId : tokenId
-  });
+  const { tokenDecimals, tokenAvatar, tokenLabel, type, esdtPrice } =
+    useGetTokenDetails({
+      tokenId: nonce && nonce?.length > 0 ? nftId : tokenId
+    });
 
   const formattedAmount = formatAmount({
     input: isTokenTransaction
@@ -95,117 +90,92 @@ export const SignStepBody = ({
     addCommas: true
   });
 
-  const extraGuardianStep = isGuarded ? 1 : 0;
-  const totalSteps = allTransactions.length + extraGuardianStep;
   const scamReport = currentTransaction.receiverScamInfo;
-  const showProgressSteps = totalSteps > 1;
   const classes = useSignStepsClasses(scamReport);
 
   const token = isNft ? nftId : tokenId || egldLabel;
   const shownAmount = isNft ? amount : formattedAmount;
 
+  const { price: egldPrice } = useGetEgldPrice();
+  let tokenPrice;
+
+  if (isEgld && egldPrice) {
+    tokenPrice = egldPrice;
+  }
+
+  if (isNft) {
+    tokenPrice = null;
+  }
+
+  if (isEsdt && type) {
+    tokenPrice = esdtPrice ?? null;
+  }
+
+  const shouldShowAmount =
+    isEgld || isEsdt || (Boolean(type) && type !== NftEnumType.NonFungibleESDT);
+
   return (
-    <>
-      {currentTransaction.transaction && (
-        <>
-          {showProgressSteps && (
-            <ProgressSteps
-              totalSteps={totalSteps}
-              currentStep={currentStep + 1} // currentStep starts at 0
-              className={classNames(globalStyles.mb4, progressClassName)}
+    <div className={styles.summary}>
+      <div className={styles.fields}>
+        {isNft && type && (
+          <NftSftPreviewComponent
+            txType={type}
+            tokenLabel={tokenLabel}
+            tokenId={tokenId}
+            tokenAvatar={tokenAvatar}
+          />
+        )}
+
+        <ConfirmReceiver
+          scamReport={scamReport}
+          receiver={
+            multiTxData
+              ? new Address(receiver).bech32()
+              : currentTransaction.transaction.getReceiver().toString()
+          }
+        />
+
+        <div className={styles.columns}>
+          {shouldShowAmount && (
+            <div className={styles.column}>
+              <ConfirmAmount
+                tokenAvatar={tokenAvatar}
+                amount={shownAmount}
+                token={token}
+                tokenType={isEgld ? egldLabel : type}
+                tokenPrice={tokenPrice}
+              />
+            </div>
+          )}
+
+          <div className={styles.column}>
+            <ConfirmFee
+              tokenAvatar={tokenAvatar}
+              egldLabel={egldLabel}
+              transaction={currentTransaction.transaction}
             />
-          )}
-
-          <div
-            data-testid='transactionTitle'
-            className={classNames(classes.formGroup, inputGroupClassName)}
-          >
-            <div className={classNames(classes.formLabel, inputLabelClassName)}>
-              To
-            </div>
-
-            <div className={inputValueClassName} data-testid='confirmReceiver'>
-              {multiTxData
-                ? new Address(receiver).bech32()
-                : currentTransaction.transaction.getReceiver().toString()}
-            </div>
-
-            {scamReport && (
-              <div
-                className={classNames(classes.scamReport, scamAlertClassName)}
-              >
-                <span>
-                  <FontAwesomeIcon
-                    icon={faExclamationTriangle}
-                    className={classes.scamReportIcon}
-                  />
-
-                  <small data-testid='confirmScamReport'>{scamReport}</small>
-                </span>
-              </div>
-            )}
           </div>
+        </div>
 
-          <div
-            className={classNames(classes.tokenWrapper, inputGroupClassName)}
-          >
-            <div
-              className={classNames(classes.tokenLabel, inputLabelClassName)}
-            >
-              Token
-            </div>
+        {currentTransaction.transaction.getData() && (
+          <TransactionData
+            isScCall={!tokenId}
+            data={currentTransaction.transaction.getData().toString()}
+            highlight={multiTxData}
+            className={inputGroupClassName}
+            innerTransactionDataClasses={{
+              transactionDataInputLabelClassName: inputLabelClassName,
+              transactionDataInputValueClassName: inputValueClassName
+            }}
+          />
+        )}
 
-            <div className={inputValueClassName} data-testid='confirmToken'>
-              <div className={classes.tokenValue}>
-                <TokenDetails.Icon tokenAvatar={tokenAvatar} token={token} />
-
-                <div className={globalStyles.mr2}></div>
-                <TokenDetails.Label token={token} />
-              </div>
-            </div>
-          </div>
-
-          <div className={inputGroupClassName}>
-            <div
-              className={classNames(
-                classes.tokenAmountLabel,
-                inputLabelClassName
-              )}
-            >
-              Amount
-            </div>
-
-            <div
-              className={classNames(
-                classes.tokenAmountValue,
-                inputValueClassName
-              )}
-              data-testid='confirmAmount'
-            >
-              {shownAmount}
-            </div>
-          </div>
-
-          {currentTransaction.transaction.getData() && (
-            <TransactionData
-              isScCall={!tokenId}
-              data={currentTransaction.transaction.getData().toString()}
-              highlight={multiTxData}
-              className={inputGroupClassName}
-              innerTransactionDataClasses={{
-                transactionDataInputLabelClassName: inputLabelClassName,
-                transactionDataInputValueClassName: inputValueClassName
-              }}
-            />
-          )}
-
-          {error && (
-            <p className={classNames(classes.errorMessage, errorClassName)}>
-              {error}
-            </p>
-          )}
-        </>
-      )}
-    </>
+        {error && (
+          <p className={classNames(classes.errorMessage, errorClassName)}>
+            {error}
+          </p>
+        )}
+      </div>
+    </div>
   );
 };