+++ src/apiCalls/index.ts
@@ -1,5 +1,6 @@
 export * from './transactions';
 export * from './accounts';
 export * from './configuration';
+export * from './economics';
 export * from './getScamAddressData';
 export * from './endpoints';