+++ src/hooks/transactions/batch/tracker/useUpdateBatch.ts
@@ -1,6 +1,6 @@
 import { useCallback } from 'react';
-import { useDispatch, useSelector } from 'reduxStore/DappProviderContext';
-import { signedTransactionsSelector } from 'reduxStore/selectors';
+import { useGetAccount } from 'hooks/account';
+import { useDispatch } from 'reduxStore/DappProviderContext';
 import { updateSignedTransactionStatus } from 'reduxStore/slices';
 import { getTransactionsDetails } from 'services/transactions/getTransactionsDetails';
 import {
@@ -9,46 +9,10 @@ import {
   TransactionServerStatusesEnum
 } from 'types';
 import { refreshAccount } from 'utils/account/refreshAccount';
-import { sequentialToFlatArray } from 'utils/transactions/batch/sequentialToFlatArray';
-import { useGetBatches } from '../useGetBatches';
 
 export function useUpdateBatch() {
   const dispatch = useDispatch();
-  const { batchTransactionsArray } = useGetBatches();
-  const signedTransactions = useSelector(signedTransactionsSelector);
-
-  const handleBatchErrors = useCallback(
-    ({
-      sessionId,
-      batchTransactions
-    }: {
-      sessionId: string;
-      batchTransactions: SignedTransactionType[];
-    }) => {
-      for (const transaction of batchTransactions) {
-        if (!signedTransactions) {
-          continue;
-        }
-
-        const signedTransaction = signedTransactions[
-          sessionId
-        ]?.transactions?.find((tx) => tx.signature === transaction.signature);
-
-        if (!signedTransaction) {
-          continue;
-        }
-
-        dispatch(
-          updateSignedTransactionStatus({
-            sessionId,
-            status: TransactionServerStatusesEnum.notExecuted,
-            transactionHash: signedTransaction.hash
-          })
-        );
-      }
-    },
-    [dispatch, signedTransactions]
-  );
+  const { address } = useGetAccount();
 
   const handleBatchSuccess = useCallback(
     ({
@@ -94,35 +58,24 @@ export function useUpdateBatch() {
 
   return useCallback(
     async (props?: {
-      batchId: string;
+      sessionId: string;
       isBatchFailed?: boolean;
       dropUnprocessedTransactions?: boolean;
       shouldRefreshBalance?: boolean;
+      transactions?: SignedTransactionType[];
     }) => {
       if (!props) {
         return;
       }
 
-      const batch = batchTransactionsArray.find(
-        (batch) => batch.batchId === props.batchId
-      );
-      if (!batch) {
+      const { transactions, isBatchFailed, sessionId } = props;
+
+      if (!transactions || transactions.length === 0) {
         return;
       }
 
-      const { batchId, transactions } = batch;
-      const sessionId = batchId.split('-')[0];
-      if (!sessionId) {
-        return;
-      }
-
-      const transactionsFlatArray = sequentialToFlatArray({ transactions });
-      if (transactionsFlatArray.length === 0) {
-        return;
-      }
-
-      if (props.isBatchFailed) {
-        for (const transaction of transactionsFlatArray) {
+      if (isBatchFailed) {
+        for (const transaction of transactions) {
           dispatch(
             updateSignedTransactionStatus({
               sessionId,
@@ -135,9 +88,7 @@ export function useUpdateBatch() {
       }
 
       const { data, success } = await getTransactionsDetails(
-        transactionsFlatArray
-          .map(({ hash }) => hash)
-          .filter((hash) => Boolean(hash))
+        transactions.map(({ hash }) => hash).filter((hash) => Boolean(hash))
       );
 
       if (success && data) {
@@ -145,12 +96,7 @@ export function useUpdateBatch() {
           sessionId,
           dropUnprocessedTransactions: props.dropUnprocessedTransactions,
           serverTransactions: data,
-          batchTransactions: transactionsFlatArray
-        });
-      } else {
-        handleBatchErrors({
-          sessionId,
-          batchTransactions: transactionsFlatArray
+          batchTransactions: transactions
         });
       }
 
@@ -158,6 +104,6 @@ export function useUpdateBatch() {
         await refreshAccount();
       }
     },
-    [dispatch, batchTransactionsArray]
+    [dispatch, address, handleBatchSuccess]
   );
 }