+++ src/hooks/transactions/useGetTokenDetails.tsx
@@ -1,8 +1,9 @@
 import axios from 'axios';
 import useSwr from 'swr';
 
-import { COLLECTIONS_ENDPOINT, TOKENS_ENDPOINT } from 'apiCalls/endpoints';
+import { NFTS_ENDPOINT, TOKENS_ENDPOINT } from 'apiCalls/endpoints';
 import { useGetNetworkConfig } from 'hooks/useGetNetworkConfig';
+import { NftEnumType } from 'types/tokens.types';
 import { getIdentifierType } from 'utils/validation/getIdentifierType';
 
 export interface TokenAssets {
@@ -16,12 +17,22 @@ export interface TokenAssets {
   lockedAccounts?: { [key: string]: string };
 }
 
+export interface TokenMediaType {
+  url?: string;
+  originalUrl?: string;
+  thumbnailUrl?: string;
+  fileType?: string;
+  fileSize?: number;
+}
+
 interface TokenOptionType {
   tokenLabel: string;
   tokenDecimals: number;
   tokenAvatar: string;
   assets?: TokenAssets;
+  type?: NftEnumType;
   error?: string;
+  esdtPrice?: number;
 }
 
 interface TokenInfoResponse {
@@ -29,7 +40,10 @@ interface TokenInfoResponse {
   name: string;
   ticker: string;
   decimals: number;
+  type?: NftEnumType;
   assets: TokenAssets;
+  media?: TokenMediaType[];
+  price: number;
 }
 
 const fetcher = (url: string) =>
@@ -41,15 +55,10 @@ export function useGetTokenDetails({
   tokenId: string;
 }): TokenOptionType {
   const { network } = useGetNetworkConfig();
+  const { isNft } = getIdentifierType(tokenId);
 
-  const { isEsdt, isNft } = getIdentifierType(tokenId);
-  const tokenEndpoint = isEsdt ? TOKENS_ENDPOINT : COLLECTIONS_ENDPOINT;
-  let tokenIdentifier = tokenId;
-
-  if (isNft) {
-    const [firstPart, secondPart] = tokenId.split('-');
-    tokenIdentifier = `${firstPart}-${secondPart}`;
-  }
+  const tokenIdentifier = tokenId;
+  const tokenEndpoint = isNft ? NFTS_ENDPOINT : TOKENS_ENDPOINT;
 
   const {
     data: selectedToken,
@@ -73,15 +82,17 @@ export function useGetTokenDetails({
     ? selectedToken?.decimals
     : Number(network.decimals);
   const tokenLabel = selectedToken ? selectedToken?.name : '';
-  const tokenAvatar = selectedToken ? `${selectedToken?.assets?.svgUrl}` : '';
-
-  const assets = selectedToken?.assets;
+  const tokenAvatar = selectedToken
+    ? selectedToken?.assets?.svgUrl ?? selectedToken?.media?.[0].thumbnailUrl
+    : '';
 
   return {
     tokenDecimals: tokenDecimals,
     tokenLabel,
+    type: selectedToken?.type,
     tokenAvatar,
-    assets,
+    assets: selectedToken?.assets,
+    esdtPrice: selectedToken?.price,
     error
   };
 }