+++ src/services/transactions/clearTransactions.ts
@@ -1,8 +1,11 @@
+import { buildBatchId } from 'hooks/transactions/helpers/buildBatchId';
+import { accountSelector } from 'reduxStore/selectors';
 import {
   clearSignedTransaction,
   clearAllTransactionsToSign,
   clearAllSignedTransactions,
-  clearBatchTransactions
+  clearBatchTransactions,
+  removeTransactionToast
 } from 'reduxStore/slices';
 import { store } from 'reduxStore/store';
 
@@ -11,6 +14,21 @@ export function removeTransactionsToSign(sessionId: string) {
 }
 export function removeSignedTransaction(sessionId: string) {
   store.dispatch(clearSignedTransaction(sessionId));
+
+  const account = accountSelector(store.getState());
+  store.dispatch(
+    clearBatchTransactions({
+      batchId: buildBatchId({
+        sessionId,
+        address: account?.address ?? ''
+      })
+    })
+  );
+}
+
+export function deleteTransactionToast(sessionId: string) {
+  store.dispatch(removeTransactionToast(sessionId));
+  removeSignedTransaction(sessionId);
 }
 
 export function removeAllSignedTransactions() {